Add pipe for reversed time field translation
Because of https://issues.apache.org/jira/browse/CASSANDRA-7867,
sorting in descending order does not always work correctly in cassandra.
A common case where this arises is when a row is intended to be ordered
in most recent-first order using a timestamp column. This adds a
workaround for this case that translates Time fields to negative
integers, so that an ascending ordering will return most recent first.

Additionally, modifies the CassandraItems and CassModKeyEnumerator pipes
to return items with indifferent access, so that field lookup by symbol
will work properly.